safari-534.53-branch/LayoutTests/svg/dom/path-totalLength-expected.txt

This tests getTotalLength of SVG path.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getTotalLength('M0,20 L400,20 L640,20') is 640
PASS getTotalLength('M0,20 L400,20 L640,20 z') is 1280
PASS getTotalLength('M0,20 L400,20 z M 320,20 L640,20') is 1120
PASS successfullyParsed is true

TEST COMPLETE
